trying to send state to components from dashboard
