+++ frontend/debug.log
@@ -6,3 +6,19 @@
 [1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1215/152300.962:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1215/163150.577:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.954:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1216/025430.966:ERROR:crash_report_database_win.cc(469)] failed to stat report
